pyproject.toml
[project]
name = "aws-by-example"
version = "0.1.0"
description = "Simple examples to test and illustrate aws behaviour."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib>=2.221.1",
    "boto3>=1.40.64",
    "jupyter>=1.1.1",
    "marimo[recommended]>=0.17.7",
    "pytest>=8.4.2",
    "pytest-xdist>=3.8.0",
]

[dependency-groups]
dev = [
    "boto3-stubs>=1.40.64",
    "ipython>=9.6.0",
    "mypy>=1.18.2",
    "nbqa>=1.9.1",
    "ruff>=0.14.3",
]

[tool.ruff]
line-length = 140

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
